(chore) moved selectors to their own folder
